# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MSV",
    "description": "",
    "version": "0.1.9"
  },
  "paths": {
...
    "/api/v1/enable-disable-customer/{user_id}": {
      "get": {
        "tags": [
          "Customer"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "type": "string",
            "description": "user id"
          }
        ],
        "summary": "enable disable customer",
        "security": [
          {
            "bearerAuth": {
              "$ref": "#/components/securitySchemes/bearerAuth"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of users"
          }
        }
      }
    },
...
    "/api/v1/forgot-password": {
      "post": {
        "tags": [
          "Customer"
        ],
        "summary": "forgot password",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A list of users"
          }
        }
      }
    },
    "/api/v1/change-password": {
      "post": {
        "tags": [
          "Customer"
        ],
        "summary": "change password",
        "security": [
          {
            "bearerAuth": {
              "$ref": "#/components/securitySchemes/bearerAuth"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "old_password": {
                    "type": "string"
                  }
                },
                "required": [
                  "password",
                  "old_password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A list of users"
          }
        }
      }
    },
    "/api/v1/change-password-mail": {
      "post": {
        "tags": [
          "Customer"
        ],
        "summary": "change password",
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "authorization": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "password",
                  "authorization"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A list of users"
          }
        }
      }
    },
...
  },
    "components": {
...
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
...
    }
  }